+++ src/rules/selector-pseudo-class-no-unknown/index.js
@@ -1,6 +1,8 @@
 import { isString } from "lodash"
 import { vendor } from "postcss"
 import {
+  isStandardRule,
+  isStandardSelector,
   parseSelector,
   report,
   ruleMessages,
@@ -29,6 +31,7 @@ export default function (actual, options) {
     if (!validOptions) { return }
 
     root.walkRules(rule => {
+      if (!isStandardRule(rule)) { return }
       const selector = rule.selector
 
       if (selector.indexOf(":") === -1) { return }
@@ -36,6 +39,8 @@ export default function (actual, options) {
       parseSelector(selector, result, rule, selectorTree => {
         selectorTree.walkPseudos(pseudoNode => {
           const pseudoClass = pseudoNode.value
+
+          if (!isStandardSelector(pseudoClass)) { return }
 
           // Ignore pseudo-elements
           if (pseudoClass.indexOf("::") !== -1) { return }