+++ src/rules/selector-pseudo-class-parentheses-space-inside/index.js
@@ -1,5 +1,6 @@
 import _ from "lodash"
 import {
+  isStandardRule,
   parseSelector,
   report,
   ruleMessages,
@@ -28,6 +29,7 @@ export default function (expectation) {
     if (!validOptions) { return }
 
     root.walkRules(rule => {
+      if (!isStandardRule(rule)) { return }
       if (rule.selector.indexOf("(") === -1) { return }
 
       parseSelector(rule.selector, result, rule, selectorTree => {